chore: merge image and images dir inside the scanner
they seem responsible for the same domain, merging them.
if we need a followup split to subdivide responsibilities, we can continue from here.
